# RFC: change optimization level to 1