--- a/src/Tokenize.jl
+++ b/src/Tokenize.jl
@@ -1,5 +1,9 @@
 module Tokenize
 
+if isdefined(Base, :Experimental) && isdefined(Base.Experimental, Symbol("@optlevel"))
+    @eval Base.Experimental.@optlevel 1
+end
+
 include("token.jl")
 include("lexer.jl")