@pfitzseb 2021-08-03
I wouldn't mind this, but 40% additional runtime for 60% saved compilation time doesn't sound super convincing if you plan on tokenizing loads of code.