@KristofferC 2021-08-04
Sure if you look at the percentages you would get that impression. My point is that looking at the absolute numbers, if you use `RawToken` you would have to lex ~30x the amount of source code as Julia itself has for it to be a benefit. I just think that is pretty unlikely in most processes.