@pfitzseb 2021-08-04
Right. I was mostly thinking about the LanguageServer case where we lex a *lot* of code (because each edit currently causes re-parsing of the whole file, of which lexing is of course only a small part).

So yeah, let's get this merged imho.