@KristofferC 2021-08-04
Yeah, I thought about vscode too but the parsing and handling of the surrounding data structures should dominate that. I can try to make a benchmark.