@pfitzseb 2021-08-04
If we include linting in the benchmark then it's ~10%:
```
Tokenize on  master [?] 
λ j1 ~/Documents/tokenize-bench.jl
  3.015 s (15077306 allocations: 974.27 MiB)

Tokenize on  kc/opt_level [?] 
λ j1 ~/Documents/tokenize-bench.jl
  3.302 s (15088796 allocations: 974.67 MiB)
```