@KristofferC 2021-08-04
I am a bit surprised by the results since I thought CSTParser used `RawToken` and that only had a 0.02s slowdown on the whole of Julia, but here the slowdowns are much larger.