@pfitzseb 2021-08-04
It does use `RawToken`s, so no idea what causes the slow down.