@ZacLN 2021-08-08
It's definitely worth it. We're parsing frequently in vscode but never that much at any one time (except on startup).