@KristofferC 2021-08-09
> I have never been able to work out why tokenize and CSTParser.parse are quite so slow on their first run. Does anyone have an idea?

From this PR it looks like quite a bit is in LLVM. Maybe we have big functions and inline a bit too much?